Rename build_api to BuildApi.js
