+++ src/js/tabs/cli.js
@@ -1,7 +1,7 @@
 import { i18n } from "../localization";
 import Clipboard from "../Clipboard";
 import GUI from '../gui';
-import BuildApi from '../build_api';
+import BuildApi from '../BuildApi';
 
 const cli = {
     lineDelayMs: 15,
